ADEN-5783 Extract ads config to separate module
