+++ app/modules/video-players/ooyala-v4.js
@@ -1,6 +1,6 @@
 import Ads from '../ads';
 import BasePlayer from './base';
-import moatVideoTracker from './moat-video-tracker';
+import OoyalaVideoAds from './ooyala-video-ads';
 import config from '../../config/environment';
 
 export const ooyalaAssets = {
@@ -52,72 +52,9 @@ export default class OoyalaV4Player extends BasePlayer {
 		window.OO.ready(() => {
 			Ads.getInstance()
 				.waitForReady()
-				.then(() => this.setupAds())
-				.then(() => window.OO.Player.create(this.containerId, this.params.videoId, this.params));
+				.then(() => (new OoyalaVideoAds(this.params)).getOoyalaConfig())
+				.then((params) => window.OO.Player.create(this.containerId, this.params.videoId, params));
 		});
-	}
-
-	setupAds() {
-		if (this.params.noAds) {
-			return null;
-		} else if (this.isA9VideoEnabled()) {
-			return this.parseBidderParameters()
-				.catch(() => {})
-				.then((additionalParams) => this.setupAdManager(additionalParams));
-		} else {
-			return this.setupAdManager();
-		}
-	}
-
-	setupAdManager(additionalParams) {
-		additionalParams = additionalParams || {};
-		this.params['google-ima-ads-manager'] = this.getAdsManagerConfig(this.buildVAST(additionalParams));
-		this.params.replayAds = false;
-	}
-
-	buildVAST(slotParams) {
-		slotParams = slotParams || {};
-		slotParams.pos = 'FEATURED';
-		slotParams.src = 'premium';
-
-		return Ads.getInstance().buildVastUrl(640 / 480, slotParams, {
-			contentSourceId: this.params.dfpContentSourceId,
-			videoId: this.params.videoId
-		});
-	}
-
-	parseBidderParameters() {
-		let a9 = Ads.getInstance().a9;
-
-		if (!a9 || !this.isA9VideoEnabled()) {
-			return {};
-		}
-
-		return a9.waitForResponse()
-			.then(() => a9.getSlotParams('FEATURED'))
-	}
-
-	isA9VideoEnabled() {
-		let ads = Ads.getInstance();
-		return Ads.getInstance().a9 && ads.currentAdsContext && ads.currentAdsContext.bidders && ads.currentAdsContext.bidders.a9Video;
-	}
-
-	getAdsManagerConfig(vastUrl) {
-		return {
-			all_ads: [
-				{
-					tag_url: vastUrl
-				}
-			],
-			useGoogleAdUI: true,
-			useGoogleCountdown: false,
-			onBeforeAdsManagerStart(IMAAdsManager) {
-				// mutes VAST ads from the very beginning
-				// FIXME with VPAID it causes volume controls to be in incorrect state
-				IMAAdsManager.setVolume(0);
-			},
-			onAdRequestSuccess: this.onAdRequestSuccess.bind(this)
-		};
 	}
 
 	/**
@@ -147,24 +84,5 @@ export default class OoyalaV4Player extends BasePlayer {
 	 */
 	playerDidLoad() {
 		this.createPlayer();
-	}
-
-	onAdRequestSuccess(IMAAdsManager, uiContainer) {
-		if (Ads.getInstance().currentAdsContext.opts.isMoatTrackingForFeaturedVideoEnabled) {
-			moatVideoTracker(IMAAdsManager, uiContainer, window.google.ima.ViewMode.NORMAL, 'ooyala', 'featured-video');
-		}
-
-		// that's a hack for autoplay on mobile for VPAID ads
-		// VPAID ads still don't work perfectly
-		let initiallyResumed = false;
-		IMAAdsManager.addEventListener('pause', (eventData) => {
-			if (eventData.getAd().getApiFramework() === 'VPAID') {
-				if (!initiallyResumed) {
-					IMAAdsManager.resume();
-					// we don't use removeEventListener because it doesn't work as expected
-					initiallyResumed = true;
-				}
-			}
-		}, false, this);
 	}
 }
